+++ ChangeLog.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - added function empty
 - added function mapv
+- added support to change 'stdout' stream by adding the variable `*out*`